# savequest-backend/swagger.yaml
openapi: 3.0.0
info:
  title: SaveQuest Plaid API
  version: 1.0.0
  description: |
    API endpoints for Plaid integration in SaveQuest backend.
    
    **Authentication:**
    - All routes (except `/plaid/webhook`) require a Bearer JWT (Firebase ID token) in the `Authorization` header.
    - Example: `Authorization: Bearer <Firebase ID token>`
    - The `/plaid/webhook` endpoint is public for Plaid to send events.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:5000
security:
  - bearerAuth: []
...
paths:
...
  /users:
    post:
      tags:
        - User Endpoints
      summary: Create a new user
      description: |
        Creates a new user in the database. Fails if a user with the same userId already exists.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - name
                - email
              properties:
                userId:
                  type: string
                  description: Unique identifier for the user.
                  example: "user_123"
                name:
                  type: string
                  description: Name of the user.
                  example: "Alice"
                email:
                  type: string
                  description: Email address of the user.
                  example: "alice@example.com"
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
                  displayName:
                    type: string
                  email:
                    type: string
                  photoURL:
                    type: string
                  providerId:
                    type: string
                  phoneNumber:
                    type: string
                  emailVerified:
                    type: boolean
                  updatedAt:
                    type: string
                    format: date-time
              examples:
                success:
                  value:
                    uid: "ZnOrsnQcdwgFO3ELvWOKmp3tUi53"
                    displayName: "Connor Li"
                    email: "connor@example.com"
                    photoURL: "https://lh3.googleusercontent.com/a-/AOh14Gg..."
                    providerId: "google.com"
                    phoneNumber: "+1234567890"
                    emailVerified: true
                    updatedAt: "2025-07-15T12:00:00Z"
        '400':
          description: Invalid input or user already exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                error:
                  value:
                    error: "User already exists."
  /users/{userId}:
    get:
      tags:
        - User Endpoints
      summary: Get user by userId
      description: |
        Retrieves a user by their userId.
      operationId: getUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: Unique identifier for the user.
      responses:
        '200':
          description: User found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
              examples:
                success:
                  value:
                    userId: "user_123"
                    name: "Alice"
                    email: "alice@example.com"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                error:
                  value:
                    error: "User not found."

    delete:
      tags:
        - User Endpoints
      summary: Delete user by userId
      description: |
        Deletes a user from the database by their userId.
      operationId: deleteUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: Unique identifier for the user.
      responses:
        '204':
          description: User deleted successfully. No content returned.
        '400':
          description: Invalid userId or request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                error:
                  value:
                    error: "userId is required."
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                error:
                  value:
                    error: "User not found."

    put:
      tags:
        - User Endpoints
      summary: Update user by userId
      description: |
        Updates an existing user's information. Fails if the user does not exist.
      operationId: updateUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: Unique identifier for the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Updated name of the user.
                  example: "Alice Smith"
                email:
                  type: string
                  description: Updated email address of the user.
                  example: "alice.smith@example.com"
      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
              examples:
                success:
                  value:
                    userId: "user_123"
                    name: "Alice Smith"
                    email: "alice.smith@example.com"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                error:
                  value:
                    error: "User not found."